webdav: removed text
Webdav is always available; it does not need to be enabled per VO or poolgroup. Only, a user should have his CUA account correctly mapped within dCache and this account should be authorized to access a dir or file.

Removed also "below are examples with curl and wget" because it's not correct anymore.