Removed fractals.c from all example tarballs
Fractals source code was present in 4 tarballs (or more? did I miss some?), which makes it very likely that you forget to update one of them in case of a fix in this source code. And I noticed, because I did a fix (see some commits ago) while updating the ToPoS example.