webdav: added --fail  to curl download
The idea of --fail is that an error message that is sent by the server does not end up instead of the file you thought you were downloading. But the option is only mentioned in the upload command. I've added it to the download command; I've not tested it though.