Create host_certificates.rst
We didn't have any information about host certificates. For basic use it's not relevant, but it becomes relevant when users start transfers from host that have not been set up by us. The webdav client page refers to the CA certificates in /etc/grid-security/certificates/, so it may be useful to elaborate a bit on this. It may also become relevant for users of the CDI.